+++ packages/gatsby-source-mongodb/src/gatsby-node.js
@@ -91,7 +91,9 @@ function createNodes(
           parent: `__${collectionName}__`,
           children: [],
           internal: {
-            type: `mongodb${caps(dbName)}${caps(collectionName)}`,
+            type: `mongodb${sanitizeName(dbName)}${sanitizeName(
+              collectionName
+            )}`,
             content: JSON.stringify(item),
             contentDigest: createContentDigest(item),
           },
@@ -135,8 +137,8 @@ function createNodes(
   })
 }
 
-function caps(s) {
-  return s.replace(/\b\w/g, l => l.toUpperCase())
+function sanitizeName(s) {
+  return s.replace(/[^_a-zA-Z0-9]/, ``).replace(/\b\w/g, l => l.toUpperCase())
 }
 
 function getConnectionExtraParams(extraParams) {