+++ packages/gatsby-transformer-remark/src/__tests__/extend-node.js
@@ -11,7 +11,12 @@ const {
 const extendNodeType = require(`../extend-node-type`)
 
 // given a set of nodes and a query, return the result of the query
-async function queryResult(nodes, fragment, { types = [] } = {}, additionalParameters) {
+async function queryResult(
+  nodes,
+  fragment,
+  { types = [] } = {},
+  additionalParameters
+) {
   const inferredFields = inferObjectStructureFromNodes({
     nodes,
     types: [...types],
@@ -23,7 +28,7 @@ async function queryResult(nodes, fragment, { types = [] } = {}, additionalParam
         get: () => null,
         set: () => null,
       },
-      getNodes: () => [],
+      getNodesByType: type => [],
       ...additionalParameters,
     },
     {
@@ -70,7 +75,13 @@ async function queryResult(nodes, fragment, { types = [] } = {}, additionalParam
   return result
 }
 
-const bootstrapTest = (label, content, query, test, additionalParameters = {}) => {
+const bootstrapTest = (
+  label,
+  content,
+  query,
+  test,
+  additionalParameters = {}
+) => {
   const node = {
     id: `whatever`,
     children: [],
@@ -82,7 +93,7 @@ const bootstrapTest = (label, content, query, test, additionalParameters = {}) =
   // Make some fake functions its expecting.
   const loadNodeContent = node => Promise.resolve(node.content)
 
-  it(label, async (done) => {
+  it(label, async done => {
     node.content = content
     const createNode = markdownNode => {
       queryResult(
@@ -96,8 +107,7 @@ const bootstrapTest = (label, content, query, test, additionalParameters = {}) =
         try {
           test(result.data.listNode[0])
           done()
-        }
-        catch(err) {
+        } catch (err) {
           done.fail(err)
         }
       })
@@ -107,20 +117,20 @@ const bootstrapTest = (label, content, query, test, additionalParameters = {}) =
     const createNodeId = jest.fn()
     createNodeId.mockReturnValue(`uuid-from-gatsby`)
     const createContentDigest = jest.fn().mockReturnValue(`contentDigest`)
-    await onCreateNode({
-      node,
-      loadNodeContent,
-      actions,
-      createNodeId,
-      createContentDigest,
-    },
-    { ...additionalParameters }
+    await onCreateNode(
+      {
+        node,
+        loadNodeContent,
+        actions,
+        createNodeId,
+        createContentDigest,
+      },
+      { ...additionalParameters }
     )
-    })
+  })
 }
 
 describe(`Excerpt is generated correctly from schema`, () => {
-
   bootstrapTest(
     `correctly loads an excerpt`,
     `---
@@ -133,7 +143,7 @@ Where oh where is my little pony?`,
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toMatch(`Where oh where is my little pony?`)
     }
@@ -150,7 +160,7 @@ date: "2017-09-18T23:19:51.246Z"
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toMatch(``)
     }
@@ -173,7 +183,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toMatch(`Where oh where is my little pony?`)
     },
@@ -196,7 +206,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(139)
     }
@@ -210,7 +220,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(46)
     }
@@ -224,10 +234,78 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
         title
     }
     `,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(50)
     }
+  )
+
+  bootstrapTest(
+    `given an html format, it correctly maps nested markdown to html`,
+    `---
+title: "my little pony"
+date: "2017-09-18T23:19:51.246Z"
+---
+
+Where oh [*where*](nick.com) **_is_** ![that pony](pony.png)?`,
+    `excerpt(format: HTML)
+    frontmatter {
+        title
+    }
+    `,
+    node => {
+      expect(node).toMatchSnapshot()
+      expect(node.excerpt).toMatch(
+        `<p>Where oh <a href="nick.com"><em>where</em></a> <strong><em>is</em></strong> <img src="pony.png" alt="that pony">?</p>`
+      )
+    }
+  )
+
+  bootstrapTest(
+    `given an html format, it prunes large excerpts`,
+    `---
+title: "my little pony"
+date: "2017-09-18T23:19:51.246Z"
+---
+
+Where oh where is that pony? Is he in the stable or down by the stream?`,
+    `excerpt(format: HTML, pruneLength: 50)
+    frontmatter {
+        title
+    }
+    `,
+    node => {
+      // expect(node).toMatchSnapshot()
+      expect(node.excerpt).toMatch(
+        `<p>Where oh where is that pony? Is he in the stable…</p>`
+      )
+    }
+  )
+
+  bootstrapTest(
+    `given an html format, it respects the excerpt_separator`,
+    `---
+title: "my little pony"
+date: "2017-09-18T23:19:51.246Z"
+---
+
+Where oh where is that pony? Is he in the stable or by the stream?
+
+<!-- end -->
+Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi auctor sit amet velit id facilisis. Nulla viverra, eros at efficitur pulvinar, lectus orci accumsan nisi, eu blandit elit nulla nec lectus. Integer porttitor imperdiet sapien. Quisque in orci sed nisi consequat aliquam. Aenean id mollis nisi. Sed auctor odio id erat facilisis venenatis. Quisque posuere faucibus libero vel fringilla.
+`,
+    `excerpt(format: HTML, pruneLength: 50)
+    frontmatter {
+        title
+    }
+    `,
+    node => {
+      expect(node).toMatchSnapshot()
+      expect(node.excerpt).toMatch(
+        `<p>Where oh where is that pony? Is he in the stable…</p>`
+      )
+    },
+    { excerpt_separator: `<!-- end -->` }
   )
 })
 
@@ -250,16 +328,15 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     frontmatter {
         title
     }`,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
-      expect(node.wordCount).toEqual(
-        {
+      expect(node.wordCount).toEqual({
         paragraphs: 2,
         sentences: 19,
         words: 150,
-        }
-      )
-    })
+      })
+    }
+  )
 
   const content = `---
 title: "my little pony"
@@ -278,16 +355,15 @@ date: "2017-09-18T23:19:51.246Z"
     frontmatter {
         title
     }`,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
-      expect(node.wordCount).toEqual(
-        {
+      expect(node.wordCount).toEqual({
         paragraphs: null,
         sentences: null,
         words: null,
-        }
-      )
-    })
+      })
+    }
+  )
 
   bootstrapTest(
     `correctly uses a default value for timeToRead`,
@@ -296,10 +372,11 @@ date: "2017-09-18T23:19:51.246Z"
     frontmatter {
         title
     }`,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.timeToRead).toBe(1)
-    })
+    }
+  )
 })
 
 describe(`Table of contents is generated correctly from schema`, () => {
@@ -324,11 +401,12 @@ some other text
     frontmatter {
         title
     }`,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(console.warn).toBeCalled()
       expect(node.tableOfContents).toBe(null)
-    })
+    }
+  )
 
   bootstrapTest(
     `correctly generates table of contents`,
@@ -352,9 +430,10 @@ final text
     frontmatter {
         title
     }`,
-  (node) => {
-    expect(node).toMatchSnapshot()
-  })
+    node => {
+      expect(node).toMatchSnapshot()
+    }
+  )
 })
 
 describe(`Links are correctly prefixed`, () => {
@@ -368,7 +447,7 @@ This is [a reference]
 [a reference]: /path/to/page2
 `,
     `html`,
-    (node) => {
+    node => {
       expect(node).toMatchSnapshot()
       expect(node.html).toMatch(`<a href="/prefix/path/to/page1">`)
       expect(node.html).toMatch(`<a href="/prefix/path/to/page2">`)