+++ packages/gatsby-transformer-sharp/src/extend-node-type.js
@@ -51,6 +51,7 @@ const fixedNodeType = ({
   getNodeAndSavePathDependency,
   reporter,
   name,
+  cache,
 }) => {
   return {
     type: new GraphQLObjectType({
@@ -80,6 +81,7 @@ const fixedNodeType = ({
                 file,
                 args,
                 reporter,
+                cache,
               })
             ).then(({ src }) => src)
           },
@@ -96,6 +98,7 @@ const fixedNodeType = ({
                 file,
                 args,
                 reporter,
+                cache,
               })
             ).then(({ srcSet }) => srcSet)
           },
@@ -151,6 +154,7 @@ const fixedNodeType = ({
           file,
           args,
           reporter,
+          cache,
         })
       ).then(o =>
         Object.assign({}, o, {
@@ -169,6 +173,7 @@ const fluidNodeType = ({
   getNodeAndSavePathDependency,
   reporter,
   name,
+  cache,
 }) => {
   return {
     type: new GraphQLObjectType({
@@ -194,6 +199,7 @@ const fluidNodeType = ({
                 file,
                 args,
                 reporter,
+                cache,
               })
             ).then(({ src }) => src)
           },
@@ -210,6 +216,7 @@ const fluidNodeType = ({
                 file,
                 args,
                 reporter,
+                cache,
               })
             ).then(({ srcSet }) => srcSet)
           },
@@ -278,6 +285,7 @@ const fluidNodeType = ({
           file,
           args,
           reporter,
+          cache,
         })
       ).then(o =>
         Object.assign({}, o, {
@@ -295,6 +303,7 @@ module.exports = ({
   pathPrefix,
   getNodeAndSavePathDependency,
   reporter,
+  cache,
 }) => {
   if (type.name !== `ImageSharp`) {
     return {}
@@ -305,6 +314,7 @@ module.exports = ({
     pathPrefix,
     getNodeAndSavePathDependency,
     reporter,
+    cache,
   }
 
   // TODO: Remove resolutionsNode and sizesNode for Gatsby v3
@@ -440,6 +450,7 @@ module.exports = ({
             resolve(
               base64({
                 file,
+                cache,
               })
             )
           } else {