+++ packages/gatsby/cache-dir/loader.js
@@ -1,6 +1,5 @@
 import pageFinderFactory from "./find-page"
 import emitter from "./emitter"
-import stripPrefix from "./strip-prefix"
 import prefetchHelper from "./prefetch"
 
 const preferDefault = m => (m && m.default) || m
@@ -16,7 +15,6 @@ let fetchingPageResourceMapPromise = null
 let fetchedPageResourceMap = false
 let apiRunner
 const failedPaths = {}
-const failedResources = {}
 const MAX_HISTORY = 5
 
 const jsonPromiseStore = {}
@@ -72,6 +70,7 @@ const fetchResource = resourceName => {
               if (req.status === 200) {
                 resolve(JSON.parse(req.responseText))
               } else {
+                delete jsonPromiseStore[resourceName]
                 reject()
               }
             }
@@ -99,10 +98,6 @@ const fetchResource = resourceName => {
           succeeded: !failed,
         })
 
-        if (!failedResources[resourceName]) {
-          failedResources[resourceName] = failed
-        }
-
         fetchHistory = fetchHistory.slice(-MAX_HISTORY)
 
         resolve(component)
@@ -112,10 +107,12 @@ const fetchResource = resourceName => {
 
 const prefetchResource = resourceName => {
   if (resourceName.slice(0, 12) === `component---`) {
-    createComponentUrls(resourceName).forEach(url => prefetchHelper(url))
+    return Promise.all(
+      createComponentUrls(resourceName).map(url => prefetchHelper(url))
+    )
   } else {
     const url = createJsonURL(jsonDataPaths[resourceName])
-    prefetchHelper(url)
+    return prefetchHelper(url)
   }
 }
 
@@ -148,14 +145,14 @@ const handleResourceLoadError = (path, message) => {
 
 const onPrefetchPathname = pathname => {
   if (!prefetchTriggered[pathname]) {
-    apiRunner(`onPrefetchPathname`, { pathname: pathname })
+    apiRunner(`onPrefetchPathname`, { pathname })
     prefetchTriggered[pathname] = true
   }
 }
 
 const onPostPrefetchPathname = pathname => {
   if (!prefetchCompleted[pathname]) {
-    apiRunner(`onPostPrefetchPathname`, { pathname: pathname })
+    apiRunner(`onPostPrefetchPathname`, { pathname })
     prefetchCompleted[pathname] = true
   }
 }
@@ -165,13 +162,6 @@ const onPostPrefetchPathname = pathname => {
 // we load all resources for likely-to-be-visited paths.
 // let pathArray = []
 // let pathCount = {}
-
-let resourcesCount = Object.create(null)
-const sortResourcesByCount = (a, b) => {
-  if (resourcesCount[a] > resourcesCount[b]) return 1
-  else if (resourcesCount[a] < resourcesCount[b]) return -1
-  else return 0
-}
 
 let findPage
 let pathScriptsCache = {}
@@ -195,14 +185,22 @@ const queue = {
   // Hovering on a link is a very strong indication the user is going to
   // click on it soon so let's start prefetching resources for this
   // pathname.
-  hovering: rawPath => {
-    const path = stripPrefix(rawPath, __PATH_PREFIX__)
+  hovering: path => {
     queue.getResourcesForPathname(path)
   },
-  enqueue: rawPath => {
-    const path = stripPrefix(rawPath, __PATH_PREFIX__)
+  enqueue: path => {
     if (!apiRunner)
       console.error(`Run setApiRunnerForLoader() before enqueing paths`)
+
+    // Skip prefetching if we know user is on slow or constrained connection
+    if (`connection` in navigator) {
+      if ((navigator.connection.effectiveType || ``).includes(`2g`)) {
+        return false
+      }
+      if (navigator.connection.saveData) {
+        return false
+      }
+    }
 
     // Tell plugins with custom prefetching logic that they should start
     // prefetching this path.
@@ -225,7 +223,7 @@ const queue = {
     ) {
       // If page wasn't found check and we didn't fetch resources map for
       // all pages, wait for fetch to complete and try find page again
-      return fetchPageResourceMap().then(() => queue.enqueue(rawPath))
+      return fetchPageResourceMap().then(() => queue.enqueue(path))
     }
 
     if (!page) {
@@ -241,12 +239,14 @@ const queue = {
 
     // Prefetch resources.
     if (process.env.NODE_ENV === `production`) {
-      prefetchResource(page.jsonName)
-      prefetchResource(page.componentChunkName)
+      Promise.all([
+        prefetchResource(page.jsonName),
+        prefetchResource(page.componentChunkName),
+      ]).then(() => {
+        // Tell plugins the path has been successfully prefetched
+        onPostPrefetchPathname(path)
+      })
     }
-
-    // Tell plugins the path has been successfully prefetched
-    onPostPrefetchPathname(path)
 
     return true
   },
@@ -290,7 +290,8 @@ const queue = {
           path,
           `Previously detected load failure for "${path}"`
         )
-        return reject()
+        reject()
+        return
       }
       const page = findPage(path)
 
@@ -303,9 +304,10 @@ const queue = {
       ) {
         // If page wasn't found check and we didn't fetch resources map for
         // all pages, wait for fetch to complete and try to get resources again
-        return fetchPageResourceMap().then(() =>
+        fetchPageResourceMap().then(() =>
           resolve(queue.getResourcesForPathname(path))
         )
+        return
       }
 
       if (!page) {
@@ -313,10 +315,12 @@ const queue = {
 
         // Preload the custom 404 page
         if (path !== `/404.html`) {
-          return resolve(queue.getResourcesForPathname(`/404.html`))
+          resolve(queue.getResourcesForPathname(`/404.html`))
+          return
         }
 
-        return resolve()
+        resolve()
+        return
       }
 
       // Use the path from the page so the pathScriptsCache uses
@@ -329,7 +333,8 @@ const queue = {
           page,
           pageResources: pathScriptsCache[path],
         })
-        return resolve(pathScriptsCache[path])
+        resolve(pathScriptsCache[path])
+        return
       }
 
       // Nope, we need to load resource(s)
@@ -352,6 +357,9 @@ const queue = {
             page,
             pageResources,
           })
+          // Tell plugins the path has been successfully prefetched
+          onPostPrefetchPathname(path)
+
           resolve(pageResources)
         })
       } else {
@@ -380,6 +388,9 @@ const queue = {
             pageResources,
           })
 
+          // Tell plugins the path has been successfully prefetched
+          onPostPrefetchPathname(path)
+
           if (doingInitialRender) {
             // We got all resources needed for first mount,
             // we can fetch resoures for all pages.
@@ -387,9 +398,6 @@ const queue = {
           }
         })
       }
-
-      // Tell plugins the path has been successfully prefetched
-      onPostPrefetchPathname(path)
     }),
 }