+++ packages/gatsby/src/internal-plugins/internal-data-bridge/gatsby-node.js
@@ -5,7 +5,7 @@ const _ = require(`lodash`)
 
 const { emitter } = require(`../../redux`)
 const { boundActionCreators } = require(`../../redux/actions`)
-const { getNode } = require(`../../redux`)
+const { getNode } = require(`../../db/nodes`)
 const createContentDigest = require(`../../utils/create-content-digest`)
 
 function transformPackageJson(json) {