+++ www/src/utils/colors.js
@@ -4,6 +4,10 @@ const colors = {
   // original palette by @SachaG
   // @see https://www.figma.com/file/J6IYJEtdRmwJQOrcZ2DfvxDD/Gatsby
   gatsby: `#663399`, // was #744c9e
+  gatsbyDark: `#442266`,
+  gatsbyDarker: `#221133`,
+  lemon: `#ffdf37`,
+  mint: `#73fff7`,
   lilac: `#8c65b3`,
   lavender: `#b190d5`,
   wisteria: `#ccb2e5`,
@@ -12,6 +16,10 @@ const colors = {
   accent: `#ffb238`, // "Mustard",
   success: `#37b635`,
   warning: `#ec1818`,
+  accentLight: `#ffeccd`,
+  accentDark: `#9e6100`,
+  skyLight: `#dcfffd`,
+  skyDark: `#0a75c2`,
   ui: {
     border: `#ede7f3`,
     bright: `#e0d6eb`,
@@ -21,9 +29,11 @@ const colors = {
   gray: {
     dark: gray(8, 270),
     copy: gray(12, 270),
+    lightCopy: gray(35, 270),
     calm: gray(46, 270),
     bright: gray(64, 270),
     light: gray(80, 270),
+    superLight: gray(96, 270),
   },
   code: {
     bg: `#fdfaf6`, // colors.a[0] #fcf6f0