+++ packages/gatsby-codemods/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.8"></a>
+
+## [1.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.7...gatsby-codemods@1.0.8) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-codemods
+
+<a name="1.0.7"></a>
+
+## [1.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.6...gatsby-codemods@1.0.7) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-codemods
+
+<a name="1.0.6"></a>
+
+## [1.0.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.5...gatsby-codemods@1.0.6) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-codemods
+
 <a name="1.0.5"></a>
 
 ## [1.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-codemods@1.0.4...gatsby-codemods@1.0.5) (2018-10-01)