+++ packages/gatsby-plugin-facebook-analytics/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2"></a>
+
+## [2.0.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-facebook-analytics@2.0.1...gatsby-plugin-facebook-analytics@2.0.2) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-plugin-facebook-analytics
+
+<a name="2.0.1"></a>
+
+## [2.0.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-facebook-analytics@2.0.0...gatsby-plugin-facebook-analytics@2.0.1) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-plugin-facebook-analytics
+
 <a name="2.0.0"></a>
 
 # [2.0.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-facebook-analytics@2.0.0-rc.1...gatsby-plugin-facebook-analytics@2.0.0) (2018-09-17)