+++ packages/gatsby-transformer-xml/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-xml/compare/gatsby-transformer-xml@2.0.4...gatsby-transformer-xml@2.0.5) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-transformer-xml
+
+<a name="2.0.4"></a>
+
+## [2.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-xml/compare/gatsby-transformer-xml@2.0.3...gatsby-transformer-xml@2.0.4) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-transformer-xml
+
+<a name="2.0.3"></a>
+
+## [2.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-xml/compare/gatsby-transformer-xml@2.0.2...gatsby-transformer-xml@2.0.3) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-transformer-xml
+
 <a name="2.0.2"></a>
 
 ## [2.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-xml/compare/gatsby-transformer-xml@2.0.1...gatsby-transformer-xml@2.0.2) (2018-10-04)